drm: Change planes is now just doing the SetProperty
Rewrite ChangePlanes() to SetChangePlanes() which is now just setting the
zpos and not doing the real commit anymore.
The related code in VideoOsdClear() and VideoOsdDrawARGB() is rewritten.
We split the original ChangePlanes() into allocating the request, setting
the properties with SetChangePlanes and committing the request.

We need this in order to do only one commit for all later, which should be
done in Frame2Display(). VideoOsdClear() and VideoOsdDrawARGB() should only
be responsible for clearing and writing to the OSD. They should mark the
OSD as active, which lets us decide in Frame2Display, which properties
for the planes should be set.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>