drm: Move OSD modesetting to Frame2Display
Depending on the new OsdShown variable which was set by VideoOsdClear()
and VideoOsdDrawARGB() we set the plane properties in Frame2Display() now.
This saves us the unnecessary atomic commit in the Clear() and DrawARGB()
functions. This is possible, because we do an atomic commit in Frame2Display()
anyway and just add the plane properties now.
Clear() and DrawARGB() just do the memcpy/memset and set some variables,
which we need later in Frame2Display().

The move of the property setting to Frame2Display() causes the properties
to be set every frame in Frame2Display().
We only need to do so, if we had any recent activity in VideoOsdClear() or
VideoOsdDrawARGB(). Introduce a dirty flag which belongs to the osd buffer,
set it in Clear/Draw and clear it after the new properties are set.

We also don't need a full modeset every frame anymore.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>