GL: Add dirty area handling
OSD is rendering faster now and we can reduce memory pressure.

Until now, openglosd doesn't make a difference, if a pixmap is dirty
or not. On fast GPU it isn't a problem, to draw every pixmap in fullsize
on every OSD flush. The smaller ARM boxes can get some performance
troubles though. To get this solved, we should only render dirty areas
of the OSD.

Before drawing the pixmaps to the osd buffer we have to find the combined
dirty area. Afterwards we loop through the layers and pixmaps as usual.
When we have a clean pixmap which intersects with the dirty area, draw the
pixmap - but only within the intersected rect. Therefore the glScissor.
Of course, draw the dirty areas, too.
We need to loop through all pixmaps to be sure to render all layers of the
affected area in ascending order.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>