Revert "video/codec: fix software decoding/deinterlacer path"
These commits introduced a problem with the threads. Dealing with
the frames in the different ringbuffers and their handling in the
different threads isn't correct and leads to SIGFPE and SIGSEGV.

This mostly occurs when a channel is switched and threads need to be
cleaned up.

The good news is that the problem for the flickering software decoded
stream on RPI4/5 is found. It is caused by the many av_frame_alloc
and *_free. Therefore the ringbuffers.

Until the introduced bugs are solved in a clean way, i'll revert these
commits.

This reverts commit 50b9d58e1ebf76a2d47284e12c3ded0c6c5e59a9.