sync: always drop and dup frames
Keep in mind, that the audio ringbuffer only can keep ~2s,
so this may have no effect in some situations, but doing nothing
at a 5s diff is weird, too.
If there is a 5s difference there might be something wrong in general.

Fix logging the number of drop/dup frames on-the-fly.