WIP: prepare SUSPENDED state
PURPOSE: The only difference to DETACHED is, that we can leave SUSPENDED not only with
an StopEvent{} but also with a PlayEvent{}.

So this enables going to suspend and "magically" wakeup up again, if some
user interaction happended.

NOTE: This does not work as expected. We can't return from SUSP with
a user interaction (channel switch...) because HasDecoder() and CanReplay()
return 0 in order to prevent an auto attach by VDR - which is intended for DETACH.
We can only return with a dedicated ATTA atm.

TODO: Let HasDecoder() and CanReplay() return true as soon as we have some user
interaction. Maybe we can take notice of a manual channel switch with cStatus?